+++ TabletopTweaks-Reworks/Config/Homebrew.cs
@@ -1,15 +1,20 @@
-﻿using TabletopTweaks.Core.Config;
+﻿using System.Collections.Generic;
+using System;
+using TabletopTweaks.Core.Config;
+using Kingmaker.Utility;
 
 namespace TabletopTweaks.Reworks.Config {
     public class Homebrew : IUpdatableSettings {
         public bool NewSettingsOffByDefault = false;
         public MythicReworkGroup MythicReworks = new MythicReworkGroup();
+        public ClassGroup Warpriest = new ClassGroup();
         public SettingGroup Feats = new SettingGroup();
         public SettingGroup MythicAbilities = new SettingGroup();
         public SettingGroup MythicFeats = new SettingGroup();
 
         public void Init() {
             MythicReworks.Init();
+            Warpriest.SetParents();
         }
 
         public void OverrideSettings(IUpdatableSettings userSettings) {
@@ -19,6 +24,8 @@ namespace TabletopTweaks.Reworks.Config {
             MythicAbilities.LoadSettingGroup(loadedSettings.MythicAbilities, NewSettingsOffByDefault);
             MythicFeats.LoadSettingGroup(loadedSettings.MythicFeats, NewSettingsOffByDefault);
             Feats.LoadSettingGroup(loadedSettings.Feats, NewSettingsOffByDefault);
+
+            Warpriest.LoadClassGroup(loadedSettings.Warpriest, NewSettingsOffByDefault);
         }
     }
     public class MythicReworkGroup : IDisableableGroup, ICollapseableGroup {
@@ -77,6 +84,42 @@ namespace TabletopTweaks.Reworks.Config {
             GoldDragon.LoadSettingGroup(group.GoldDragon, frozen);
             Legend.LoadSettingGroup(group.Legend, frozen);
             Swarm.LoadSettingGroup(group.Swarm, frozen);
+        }
+    }
+
+    public class ClassGroup : IDisableableGroup, ICollapseableGroup {
+        private bool IsExpanded = true;
+        public bool DisableAll = false;
+        public bool GroupIsDisabled() => DisableAll;
+        public void SetGroupDisabled(bool value) => DisableAll = value;
+        public NestedSettingGroup Base;
+        public SortedDictionary<string, NestedSettingGroup> Archetypes = new SortedDictionary<string, NestedSettingGroup>(StringComparer.InvariantCulture);
+
+        public ClassGroup() {
+            Base = new NestedSettingGroup(this);
+        }
+
+        public void SetParents() {
+            Base.Parent = this;
+            Archetypes.ForEach(entry => entry.Value.Parent = this);
+        }
+
+        public void LoadClassGroup(ClassGroup group, bool frozen) {
+            DisableAll = group.DisableAll;
+            Base.LoadSettingGroup(group.Base, frozen);
+            group.Archetypes.ForEach(entry => {
+                if (Archetypes.ContainsKey(entry.Key)) {
+                    Archetypes[entry.Key].LoadSettingGroup(entry.Value, frozen);
+                }
+            });
+        }
+
+        ref bool ICollapseableGroup.IsExpanded() {
+            return ref IsExpanded;
+        }
+
+        public void SetExpanded(bool value) {
+            IsExpanded = value;
         }
     }
 }